+++ h2o-algos/src/main/java/hex/tree/SharedTreeModel.java
@@ -48,9 +48,9 @@ public abstract class SharedTreeModel<M extends SharedTreeModel<M,P,O>, P extend
 
     public int _score_interval = 4000; //Adding this parameter to take away the hard coded value of 4000 for scoring each iteration every 4 secs
 
-    public float _sample_rate = 0.632f; //fraction of rows to sample for each tree
+    public double _sample_rate = 0.632; //fraction of rows to sample for each tree
 
-    public float[] _sample_rate_per_class; //fraction of rows to sample for each tree, per class
+    public double[] _sample_rate_per_class; //fraction of rows to sample for each tree, per class
 
     @Override public long progressUnits() { return _ntrees; }
 
@@ -58,8 +58,8 @@ public abstract class SharedTreeModel<M extends SharedTreeModel<M,P,O>, P extend
       return _seed == -1 ? (_seed = RandomUtils.getRNG(System.nanoTime()).nextLong()) : _seed;
     }
 
-    public float _col_sample_rate_change_per_level = 1.0f; //relative change of the column sampling rate for every level
-    public float _col_sample_rate_per_tree = 1.0f; //fraction of columns to sample for each tree
+    public double _col_sample_rate_change_per_level = 1.0f; //relative change of the column sampling rate for every level
+    public double _col_sample_rate_per_tree = 1.0f; //fraction of columns to sample for each tree
 
     /** Fields which can NOT be modified if checkpoint is specified.
      * FIXME: should be defined in Schema API annotation
